Moved more submit code into SubmitPlan classes
- Made sure that tests and pre-commit hooks pass for this commit
- Unfroze SubmitPlan class so that functions can work with one object rather than passing intermediary variables between functions before finally creating the object in build_submit_plan
- Create a shared Stage class that BayesEoRStage and pyuvsimStage can inherit from, to unify names of job steps
- Brought more Slurm functions into common module
- Fixed test for bayeseor_cli
