+++ src/valska/external_tools/bayeseor/submit.py
@@ -4,18 +4,20 @@ from __future__ import annotations
 
 import json
 import re
-import subprocess
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Any, Literal
 
 from valska.external_tools.common.submit import (
     InvalidArgumentError,
     MissingDependencyError,
-    SbatchError,
-    SubmissionError,
+    Stage,
     SubmitPlan,
-    load_manifest,
+    _jobs_path,
+    ensure_script_exists,
+    get_path_from_artefacts,
+    normalise_path,
+    run_sbatch,
 )
 from valska.external_tools.common.utils import utc_now_iso
 
@@ -23,7 +25,7 @@ _STAGE = Literal["cpu", "gpu", "all"]
 _HYP = Literal["signal_fit", "no_signal", "both"]
 _RECORD = Literal["jobs.json", "manifest"]
 
-_JOBID_RE = re.compile(r"Submitted\s+batch\s+job\s+(\d+)\s*$", re.IGNORECASE)
+
 _AFTEROK_RE = re.compile(r"afterok:(\d+)")
 _CPU_MATRIX_MARKERS: dict[str, tuple[str, ...]] = {
     "Ninv": ("Ninv.h5", "Ninv.npz"),
@@ -31,149 +33,51 @@ _CPU_MATRIX_MARKERS: dict[str, tuple[str, ...]] = {
 }
 
 
-@dataclass(frozen=True)
+class BayesEoRStage(Stage):
+    CPU = "cpu_precompute"
+    GPU = "gpu"
+    # ALL = "all"
+
+
+@dataclass()
 class BayesEoRSubmitPlan(SubmitPlan):
     """Resolved paths needed to submit a prepared BayesEoR run."""
 
-    cpu_script: Path
-    gpu_signal_fit_script: Path | None
-    gpu_no_signal_script: Path | None
-    cpu_precompute_driver_hypothesis: str | None
+    cpu_script: Path = field(init=False)
+    gpu_signal_fit_script: Path | None = field(init=False)
+    gpu_no_signal_script: Path | None = field(init=False)
+    cpu_precompute_driver_hypothesis: str | None = field(init=False)
 
+    def __post_init__(self):
+        super().__post_init__()
 
-def build_submit_plan(run_dir: Path) -> BayesEoRSubmitPlan:
-    """
-    Create a submission plan from an existing prepared run_dir.
+        artefacts = self.load_manifest().get("artefacts", {})
+        if not isinstance(artefacts, dict):
+            raise MissingDependencyError(
+                "manifest['artefacts'] is missing or not a dict"
+            )
 
-    This reads manifest.json and uses manifest['artefacts'] paths rather than guessing filenames.
-    """
-    run_dir = Path(run_dir).expanduser().resolve()
-    manifest_path = run_dir / "manifest.json"
-    manifest = load_manifest(run_dir)
-
-    artefacts = manifest.get("artefacts", {})
-    if not isinstance(artefacts, dict):
-        raise MissingDependencyError(
-            "manifest['artefacts'] is missing or not a dict"
+        cpu_script = get_path_from_artefacts(
+            artefacts, "submit_sh_cpu_precompute", required=True
+        )
+        gpu_signal = get_path_from_artefacts(
+            artefacts, "submit_sh_signal_fit_gpu_run", required=False
+        )
+        gpu_nosig = get_path_from_artefacts(
+            artefacts, "submit_sh_no_signal_gpu_run", required=False
         )
 
-    def _get_path(key: str, required: bool = True) -> Path | None:
-        p = artefacts.get(key)
-        if p is None:
-            if required:
-                raise MissingDependencyError(
-                    f"manifest artefact missing required key: {key}"
-                )
-            return None
-        return Path(str(p)).expanduser()
-
-    cpu_script = _get_path("submit_sh_cpu_precompute", required=True)
-    gpu_signal = _get_path("submit_sh_signal_fit_gpu_run", required=False)
-    gpu_nosig = _get_path("submit_sh_no_signal_gpu_run", required=False)
-
-    cpu_driver = (
-        manifest.get("bayeseor", {}).get(
-            "cpu_precompute_driver_hypothesis", None
-        )
-        if isinstance(manifest.get("bayeseor", {}), dict)
-        else None
-    )
-
-    def _normalise(p: Path | None) -> Path | None:
-        if p is None:
-            return None
-        return (run_dir / p).resolve() if not p.is_absolute() else p.resolve()
-
-    cpu_script = _normalise(cpu_script)  # type: ignore[assignment]
-    gpu_signal = _normalise(gpu_signal)
-    gpu_nosig = _normalise(gpu_nosig)
-
-    return BayesEoRSubmitPlan(
-        run_dir=run_dir,
-        manifest_path=manifest_path,
-        cpu_script=cpu_script,  # type: ignore[arg-type]
-        gpu_signal_fit_script=gpu_signal,
-        gpu_no_signal_script=gpu_nosig,
-        cpu_precompute_driver_hypothesis=str(cpu_driver)
-        if cpu_driver is not None
-        else None,
-    )
-
-
-def _ensure_script_exists(path: Path, label: str) -> None:
-    if not path.exists():
-        raise MissingDependencyError(f"Missing {label} script: {path}")
-    if not path.is_file():
-        raise MissingDependencyError(f"{label} script is not a file: {path}")
-
-
-def _run_sbatch(
-    script: Path,
-    *,
-    dependency_afterok: str | None = None,
-    sbatch_exe: str = "sbatch",
-    cwd: Path | None = None,
-    dry_run: bool = False,
-) -> tuple[str | None, str]:
-    """
-    Returns (job_id, command_str). If dry_run=True, job_id is None.
-    """
-    cmd: list[str] = [sbatch_exe]
-    if dependency_afterok:
-        cmd.append(f"--dependency=afterok:{dependency_afterok}")
-    cmd.append(str(script))
-
-    cmd_str = " ".join(cmd)
-
-    if dry_run:
-        return None, cmd_str
-
-    proc = subprocess.run(
-        cmd,
-        cwd=str(cwd) if cwd is not None else None,
-        capture_output=True,
-        text=True,
-    )
-
-    out = (proc.stdout or "").strip()
-    err = (proc.stderr or "").strip()
-
-    if proc.returncode != 0:
-        raise SbatchError(
-            "sbatch failed.\n"
-            f"Command: {cmd_str}\n"
-            f"Return code: {proc.returncode}\n"
-            f"stdout:\n{out}\n"
-            f"stderr:\n{err}\n"
+        self.cpu_precompute_driver_hypothesis = (
+            self.load_manifest()
+            .get("bayeseor", {})
+            .get("cpu_precompute_driver_hypothesis", None)
+            if isinstance(self.load_manifest().get("bayeseor", {}), dict)
+            else None
         )
 
-    m = _JOBID_RE.search(out)
-    if not m:
-        raise SbatchError(
-            "Could not parse job id from sbatch stdout.\n"
-            f"Command: {cmd_str}\n"
-            f"stdout:\n{out}\n"
-            f"stderr:\n{err}\n"
-        )
-
-    return m.group(1), cmd_str
-
-
-def _jobs_path(run_dir: Path) -> Path:
-    return run_dir / "jobs.json"
-
-
-def load_jobs(run_dir: Path) -> dict[str, Any] | None:
-    p = _jobs_path(run_dir)
-    if not p.exists():
-        return None
-    try:
-        data = json.loads(p.read_text(encoding="utf-8"))
-        return data if isinstance(data, dict) else None
-    except Exception as e:  # pragma: no cover
-        raise SubmissionError(
-            f"Failed to parse existing jobs.json: {p}\n{e}"
-        ) from e
+        self.cpu_script = normalise_path(self.run_dir, cpu_script)  # type: ignore[assignment]
+        self.gpu_signal_fit_script = normalise_path(self.run_dir, gpu_signal)
+        self.gpu_no_signal_script = normalise_path(self.run_dir, gpu_nosig)
 
 
 def _safe_int_jobid(x: Any) -> str | None:
@@ -241,71 +145,6 @@ def _find_completed_cpu_precompute_matrix_dir(run_dir: Path) -> Path | None:
     return max(common_dirs, key=lambda p: p.stat().st_mtime)
 
 
-def _merge_jobs_record(
-    existing: dict[str, Any] | None, new_result: dict[str, Any]
-) -> dict[str, Any]:
-    """
-    Merge a new submission result into an existing jobs.json record.
-
-    - Keeps stable top-level metadata (run_dir, manifest)
-    - Updates "jobs" by stage (cpu_precompute, gpu)
-    - Appends to "history" so we don't lose what happened
-    """
-    merged: dict[str, Any] = {}
-
-    # Start from existing, then overlay stable fields from new_result
-    if isinstance(existing, dict):
-        merged.update(existing)
-
-    # Always set these from new_result to reflect latest invocation context
-    for k in ("run_dir", "manifest"):
-        if k in new_result:
-            merged[k] = new_result[k]
-
-    merged["sbatch"] = new_result.get("sbatch", merged.get("sbatch", "sbatch"))
-    merged["dry_run"] = bool(new_result.get("dry_run", False))
-
-    # Keep a full submission history (append-only)
-    hist = merged.get("history")
-    if not isinstance(hist, list):
-        hist = []
-    # store a compact record (not including any existing "history")
-    hist_entry = {k: v for k, v in new_result.items() if k != "history"}
-    hist.append(hist_entry)
-    merged["history"] = hist
-
-    # Merge jobs by stage
-    merged_jobs = merged.get("jobs")
-    if not isinstance(merged_jobs, dict):
-        merged_jobs = {}
-
-    new_jobs = new_result.get("jobs")
-    if isinstance(new_jobs, dict):
-        # CPU stage record
-        cpu = new_jobs.get("cpu_precompute")
-        if isinstance(cpu, dict):
-            merged_jobs["cpu_precompute"] = cpu
-
-        # GPU stage record
-        gpu = new_jobs.get("gpu")
-        if isinstance(gpu, dict):
-            merged_jobs["gpu"] = gpu
-
-    merged["jobs"] = merged_jobs
-
-    # Keep latest submitted timestamp and commands for convenience
-    merged["submitted_utc"] = new_result.get(
-        "submitted_utc", merged.get("submitted_utc")
-    )
-    merged["stage"] = new_result.get("stage", merged.get("stage"))
-    merged["hypothesis"] = new_result.get(
-        "hypothesis", merged.get("hypothesis")
-    )
-    merged["commands"] = new_result.get("commands", merged.get("commands", []))
-
-    return merged
-
-
 def write_jobs(run_dir: Path, jobs: dict[str, Any]) -> Path:
     p = _jobs_path(run_dir)
     p.write_text(json.dumps(jobs, indent=2), encoding="utf-8")
@@ -359,10 +198,13 @@ def submit_bayeseor_run(
 
     We also keep a submission 'history' list so previous job ids are not lost.
     """
-    plan = build_submit_plan(run_dir)
 
-    _ensure_script_exists(plan.manifest_path, "manifest.json")
-    _ensure_script_exists(plan.cpu_script, "CPU precompute")
+    plan = BayesEoRSubmitPlan(
+        run_dir.expanduser().resolve(), BayesEoRStage.CPU
+    )
+
+    ensure_script_exists(plan.manifest_path, "manifest.json")
+    ensure_script_exists(plan.cpu_script, "CPU precompute")
 
     if record == "manifest":
         raise InvalidArgumentError(
@@ -370,7 +212,7 @@ def submit_bayeseor_run(
             "Use record='jobs.json' (default)."
         )
 
-    existing_jobs = load_jobs(plan.run_dir)
+    existing_jobs = plan.load_jobs()
 
     # Second-line guardrail: refuse re-submission of a stage if that stage is already recorded,
     # unless force=True. (cli_submit.py also enforces guardrails; this protects programmatic calls.)
@@ -420,7 +262,7 @@ def submit_bayeseor_run(
     # CPU submission
     # --------------------
     if stage in ("cpu", "all"):
-        jobid, cmd = _run_sbatch(
+        jobid, cmd = run_sbatch(
             plan.cpu_script,
             dependency_afterok=None,
             sbatch_exe=sbatch_exe,
@@ -493,8 +335,8 @@ def submit_bayeseor_run(
                     "manifest does not contain a signal_fit GPU submit script artefact "
                     "(submit_sh_signal_fit_gpu_run)."
                 )
-            _ensure_script_exists(plan.gpu_signal_fit_script, "signal_fit GPU")
-            jobid, cmd = _run_sbatch(
+            ensure_script_exists(plan.gpu_signal_fit_script, "signal_fit GPU")
+            jobid, cmd = run_sbatch(
                 plan.gpu_signal_fit_script,
                 dependency_afterok=dep,
                 sbatch_exe=sbatch_exe,
@@ -513,8 +355,8 @@ def submit_bayeseor_run(
                     "manifest does not contain a no_signal GPU submit script artefact "
                     "(submit_sh_no_signal_gpu_run)."
                 )
-            _ensure_script_exists(plan.gpu_no_signal_script, "no_signal GPU")
-            jobid, cmd = _run_sbatch(
+            ensure_script_exists(plan.gpu_no_signal_script, "no_signal GPU")
+            jobid, cmd = run_sbatch(
                 plan.gpu_no_signal_script,
                 dependency_afterok=dep,
                 sbatch_exe=sbatch_exe,
@@ -533,7 +375,7 @@ def submit_bayeseor_run(
     # Record jobs.json (MERGE, do not overwrite)
     # --------------------
     if not dry_run:
-        merged = _merge_jobs_record(existing_jobs, result)
+        merged = plan.merge_jobs_record(result)
         write_jobs(plan.run_dir, merged)
         # Return the merged record to reflect what is now on disk
         return merged
